move components for better folder structure
