+++ web-app/src/app/app/agents/[agentId]/jobs/components/jobs-table.tsx
